fix:系统菜单增加编辑功能、删除条件限制、添加编辑采用iframe表单形式

+++ src/main/java/com/tangerineSpecter/oms/system/service/system/MenuSettingService.java
@@ -1,6 +1,6 @@
 package com.tangerinespecter.oms.system.service.system;
 
-import com.alibaba.fastjson.JSON;
+import com.tangerinespecter.oms.common.constant.RetCode;
 import com.tangerinespecter.oms.common.result.ServiceResult;
 import com.tangerinespecter.oms.system.domain.entity.SystemMenu;
 import com.tangerinespecter.oms.system.domain.vo.SystemMenuInfoVo;
@@ -25,6 +25,14 @@ public class MenuSettingService {
         if (id == null) {
             return ServiceResult.paramError();
         }
+        SystemMenu systemMenu = systemMenuMapper.selectById(id);
+        if (systemMenu == null) {
+            return ServiceResult.error(RetCode.SYSTEM_MENU_NOT_EXIST);
+        }
+        List<SystemMenu> menuList = systemMenuMapper.selectByPid(systemMenu.getId());
+        if (menuList.size() > 0) {
+            return ServiceResult.error(RetCode.SYSTEM_MENU_CHILD_EXIST);
+        }
         systemMenuMapper.deleteById(id);
         return ServiceResult.success();
     }
@@ -34,6 +42,25 @@ public class MenuSettingService {
                 .icon("fa " + vo.getIcon()).level(vo.getLevel()).pid(vo.getPid())
                 .target(vo.getTarget()).sort(vo.getSort()).build();
         systemMenuMapper.insert(systemMenu);
+        return ServiceResult.success();
+    }
+
+    public ServiceResult detailInfo(Long id) {
+        if (id == null) {
+            return ServiceResult.paramError();
+        }
+        SystemMenu systemMenu = systemMenuMapper.selectById(id);
+        return ServiceResult.success(systemMenu);
+    }
+
+    public ServiceResult updateInfo(SystemMenuInfoVo vo) {
+        if (vo.getId() == null) {
+            return ServiceResult.paramError();
+        }
+        SystemMenu systemMenu = SystemMenu.builder().id(vo.getId()).title(vo.getTitle()).href(vo.getHref())
+                .pid(vo.getPid()).level(vo.getLevel()).sort(vo.getSort()).target(vo.getTarget())
+                .icon("fa " + vo.getIcon()).build();
+        systemMenuMapper.updateById(systemMenu);
         return ServiceResult.success();
     }
 }
